Profile page displays profile based on username
New route is /profile/:username.
Profile page requests from the server account info based on username.
Profile page is readonly and accessible for anyone but files require login.
Profile uses router-link instead of <a> so that it updates based on username
and other components can respond and the url hash used by Vue is abstracted.

Fileview requests file based on hash (/file/:hash).
The user can now update the hash on the url directly to request a new file.

Hardcode profile link to /profile/TheStruggle to show that it works.